Check dirs permissions only when not in container
