+++ cmapi/cmapi_server/__main__.py
@@ -29,7 +29,7 @@ from cmapi_server.failover_agent import FailoverAgent
 from cmapi_server.managers.application import AppManager
 from cmapi_server.managers.process import MCSProcessManager
 from cmapi_server.managers.certificate import CertificateManager
-from cmapi_server.state_checks import run_state_checks
+from cmapi_server.invariant_checks import run_invariant_checks
 from failover.node_monitor import NodeMonitor
 from mcs_node_control.models.dbrm_socket import SOCK_TIMEOUT, DBRMSocketHandler
 from mcs_node_control.models.node_config import NodeConfig
@@ -153,7 +153,7 @@ if __name__ == '__main__':
     CertificateManager.renew_certificate()
 
     # Run checks, if some of them fail -- log and exit
-    run_state_checks()
+    run_invariant_checks()
 
     app = cherrypy.tree.mount(root=None, config=CMAPI_CONF_PATH)
     root_config = {
