+++ cmapi/requirements.in
@@ -18,6 +18,6 @@ Routes==2.5.1
 typer==0.15.2
 pydantic==2.11.7
 sentry-sdk==2.34.1
-# State checks
-mr_kot==0.8.1
+# Invariant checks
+mr_kot==0.8.3
 mr_kot_fs_validators==0.1.0
